Fix mysql config permissions for Windows users
Windows docker mounts are in world writable mode, which makes mysql
ignore our port configuration. This patch solves it in a brutal way, by
manually setting the permissions of the file rather than playing with
docker mount syntaxes that may or may not work depending on the FS
driver.

Fixes #240

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>